Reworked the previous atrocious tooltip system with something sane
This was the most prominent demonstration of tech debt and code rot in this program by far.

There is one downside with the rework, and that is that the tooltip does not reset the counter if it switches between two different buttons, but this is actually pleasant in a couple ways. I know a way to fix it, but I won't.

I could have made `tip.text` a part of the panel_button function parameters, but I feel like I already have enough of those, and this way I limit the parameters of future functions as well